requirements.txt
pandas
numpy
pyyaml
boto3
sqlalchemy
psycopg2-binary
fastapi
uvicorn